Refactor interview system spec
- spec handles creation, review and cancellation of an interview
- remove create interview system spec
